+++ .github/workflows/build.yml
@@ -36,4 +36,4 @@ jobs:
         run: npm run schema postgres root
 
       - name: Test Backend-Server
-        run: npm run test -- --reset true
+        run: npm run test -- --reset true --simultaneous 16
