+++ .github/workflows/typos.yml
@@ -1,0 +1,14 @@
+name: typos
+on:
+  pull_request:
+
+jobs:
+  typos:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout Actions Repository
+        uses: actions/checkout@v4
+
+      - uses: crate-ci/typos@master
+        with:
+          config: ./typos.toml
