+++ nest-cli.json
@@ -1,0 +1,9 @@
+{
+  "$schema": "https://json.schemastore.org/nest-cli",
+  "collection": "@nestjs/schematics",
+  "sourceRoot": "src",
+  "entryFile": "executable/server",
+  "compilerOptions": {
+    "deleteOutDir": true
+  }
+}
