+++ src/api/structures/common/IRecordMerge.ts
@@ -10,7 +10,7 @@ import { tags } from "typia";
  * {@link IRecordMerge.absorbed} records.
  *
  * If there're some dependent tables of the target `table` having
- * unique constraint on foriegn key column, such dependent tables
+ * unique constraint on foreign key column, such dependent tables
  * also perform the merge process, too.
  *
  * Of course, if there're another dependent tables under those
